requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0

# RAG & Vector Store
chromadb>=0.4.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Image Processing
Pillow>=10.0.0

# Database (optional for future use)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
